+++ src/documents/query.rs
@@ -66,14 +66,14 @@ pub fn query(
         ..Default::default()
     };
 
-    let mut resp = auth
+    let resp = auth
         .client()
         .post(&url)
         .bearer_auth(auth.access_token().to_owned())
         .json(&query_request)
         .send()?;
 
-    extract_google_api_error(&mut resp, || collection_id.to_owned())?;
+    let resp = extract_google_api_error(resp, || collection_id.to_owned())?;
 
     let json: Option<Vec<dto::RunQueryResponse>> = resp.json()?;